==> R/00_all.qmd <==
---
title: "00_all"
format: html
editor: visual
---

# Introduction

Neoadjuvant chemotherapy (NACT) is gaining popularity in treating breast cancer. Achieving a pathological complete response (pCR) is seen as a positive indicator of disease-specific survival. This research "Circulating Exosomal microRNAs as Predictive Biomarkers of Neoadjuvant Chemotherapy Response in Breast Cancer" by Todorova et al. explores the potential of circulating exosomal microRNAs to forecast pCR in breast cancer patients undergoing NACT.

For our project, we used the above-mentioned article data set. The data involved collecting plasma samples from 20 patients before and after the initial treatment cycle. The samples were sequenced using an miRNA library and afterwards the reads where then mapped to an miRNA database, and this collection of gene read counts is what we will base our analysis on.

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 05_analysis_1.qmd >}}

{{< include 06_analysis_2.qmd >}}

# Function used for analysis II

```{r}
# Function to create and save the graph of a community to a PNG file
create_and_save_graph_for_community <- function(community_data, type, subcommunity, seed, output_dir) {
  i <- subcommunity
  current_graph <- community_data$communities[[i]]
  
  # Plot the graph
  g <- ggraph(current_graph) +
    geom_edge_link() +
    geom_node_point(
      aes(color = seed), 
      size = 5) +
    geom_node_text(
      aes(label = name), 
      size = 10, 
      repel = TRUE) +
    ggtitle(str_c("Community ", i, " ", type)) +
    labs(color = "Seed Nodes")
  
  # Save the graph as a PNG file
  graph_filename <- file.path(
    output_dir, 
    str_c(type, "_community_", i, ".png"))
  ggsave(
    graph_filename, 
    g, 
    width = 10, 
    height = 8)
}
```

# Conclusion

In conclusion, we found that the genes from the pCR group were more differentially expressed than the non pCR which makes sense as the pCR group were the ones responding well to the treatment.

After performing overrepresentation enrichment on our pCR and non-pCR communities we found that for the pCR complexes the following were enriched:

-   Cell metabolism: 3 groups enriched for metabolic processes: ribonucleotides, peptides.

-   RNA/DNA replication, cell cycle division control: 4 involved in replication and transcription and cell cycle

-   Immunity: 1 with enrichment for immune system pathways

-   Epigenetics: 1 for histone modification.

For the non-pCR the following were enriched:

-   Cell metabolism: 3 groups enriched for metabolic processes containing ribonucleotides and peptides. 1 mitogenic signalling

-   Cell cycle division control: 1 group involved in r and cell cycle

-   RNA/DNA replication: 1 group involved in replication and transcription

For further analysis we would look into the genes within the enriched communities and see how they may relate to the treatment of cancer patients. If we can use the expression of certain genes to determine if the patient has achieved pathological complete response to the treatment, at an early stage then we can decide whether the patient needs surgery and also if they do not achieve pCR then we can switch treatment and hopefully save them.
