==> R/02_clean.qmd <==
---
title: "02_clean"
format: html
editor: visual
---

# Cleaning the data

```{r}
#| warning: false
library("tidyverse")
```

First we explore the data

```{r}
reads_data |> 
  sample_n(10)
```

**Columns:**

1.  **Name** which include the gene names.
2.  **Identifier** which includes a link to the ENSG identifier for the gene.
3.  Then 40 columns, one for each sample. They start with 20756_0XX_genes - Total counts. Where XX is the number of the sample

So we need to clean the data and change it into tidy data meaning:

Columns = variables

Rows = Observations

Each cell = values.

## Tidying the data

```{r}
#| label: Tidying data
reads_data_long <- reads_data |> 
  
  #Pivot longer to make into tidy data set.
  pivot_longer(cols = starts_with("20756"),
               names_to = "Sample",
               values_to = "Count") |>
  
  # Change to sample number
  mutate(Sample = as.numeric(str_sub(Sample, 
                                     start = 7, 
                                     end = 9))) |>  
  #Rename Name column 
  rename("Gene"="Name")
```

Now we need to join the meta data with the count data. Since the samples and ID's match 1:1 we simply need to do a full join.

## Joining meta data and count data

```{r}
#| label: Metadata joined to the count data

# Connect the expression dataset Sample with the metadata set ID
full_reads_data <- full_join(x = reads_data_long, 
                                   y = meta_data, 
                                   join_by(Sample == ID))

write_csv(full_reads_data, 
          "../data/full_reads_data.csv")
```
