==> R/04_describe.qmd <==
---
title: "04_describe"
format: html
editor: visual
---

# Description of data

## Load libraries

```{r}
#|echo: false
library(tidyverse)
library(patchwork)
```

## Metadata overview

Our first figure represents an overview of the meta_data. We have 20 breast cancer patients before (pre_treatment) and after (post_treatment) the NACT treatment. It means that we used 40 samples for our analysis and every patient has 1 pre- and 1 post-treatment sample. The plot shows the 40 samples based on their pCR_status.

```{r, fig.height=7.5, fig.width=5.4}
#| label: fig-metadata
#| warning: false
figure1 <- meta_data |>
  ggplot(aes(x = pCR_status,
             y = ID,
             color = subject_ID))+
  geom_point(aes(shape = Treatment))+
  geom_hline(yintercept = 0)+
  coord_fixed(0.1)+
  scale_y_continuous(breaks = 1:40) +
  scale_color_discrete(labels = c("patient_01", "patient_02", "patient_03",
                                  "patient_04", "patient_05", "patient_06",
                                  "patient_07", "patient_08", "patient_09",
                                  "patient_10", "patient_11", "patient_12",
                                  "patient_13", "patient_14", "patient_15",
                                  "patient_16", "patient_17", "patient_18",
                                  "patient_19", "patient_20" ))+
  theme_minimal(base_size = 10)+
  theme(plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_text(vjust = 7,
                                   hjust = 0.5))+
  labs(x = "pCR Status",
       y ="Samples",
       title = "Samples based on pCR status by Treatment",
       color = "Number of Patients")

plot(figure1)


ggsave("../results/figure_01.jpg",
       plot = figure1,
       width = 5.4,
       height = 7.5,
       device = "jpeg")

```

## Analysis of the full data set

The full_reads_data contains all the observations and variables that we need to use for the analysis.

```{r}
full_reads_data|>
  sample_n(10)
```

To represent this data set is not that easy because the data is huge. So we arbitrarily created some plots to be able to show what our data looks like.

In this figure, we illustrate the average read counts of all the samples by the treatment.

```{r, fig.height=6, fig.width=10}
#| echo: false
figure2 <- full_reads_data |>
  group_by(Sample) |>
  summarise(av = mean(Count))|>
  mutate(Treatment = case_when(Sample %% 2 == 0 ~ "Post-treatment",
                               Sample %% 2 !=0 ~ "Pre-treatment")) |># we created the treatment column 
  ggplot(aes(x = Sample,
             y = av))+
  geom_hline(yintercept = 0)+
  scale_x_continuous(breaks = 1:40) +
  geom_col(aes(fill= Treatment),
           position = position_dodge(preserve = "single"),
           colour = "black",
           alpha = 0.5)+
  theme_minimal(base_size = 10)+
  theme(plot.margin = unit(c(1, 0.5, 0.5, 1), "cm"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_text(vjust = 7,
                                   hjust = 0.4))+
  labs(x = "Samples",
       y = "Average Read Counts",
       title = "The Average Read Counts of all the Samples by the Treatment",
       legend = "Treatment")

print(figure2)

ggsave("../results/figure_02.jpg",
       plot = figure2,
       width = 10,
       height = 6,
       device = "jpeg")
```

The 3. figure displays the differences in read counts between the pre-and post-treatment by the genes. As an example, we arbitrary chose 1-1 pCR-patient from each treatment type and filtered the read counts.

```{r, fig.height=10, fig.width=20}
#| echo: false
p1 <- full_reads_data |>
  filter(pCR_status == "pCR" &
         subject_ID == "subject_03" &
         Count > 500 &
         Treatment == "pre_treatment") |>
  ggplot(aes(x = Gene,
             y = Count,
             fill = Gene))+
  geom_hline(yintercept = 0)+
  geom_col(position = position_dodge(preserve = "single"),
           colour = "black",
           alpha = 0.5)+
  scale_y_continuous(breaks = seq(500,
                                  19000,
                                  by = 2000))+
  theme_minimal(base_size = 20)+
  theme(plot.margin = unit(c(0.5, 0.5, 0.5, 0.5),"cm"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_blank())+
  theme(legend.position = "none")+
  labs(x ="Genes",
       y ="Total Read Counts",
       title ="A Pre-treatment Sample of pCR Patient_03 filterd by 500 > Counts")

p2 <- full_reads_data |>
  filter(pCR_status == "pCR" &
         subject_ID == "subject_03" &
         Count > 500 &
         Treatment == "post_treatment") |>
  ggplot(aes(x = Gene,
             y = Count,
             fill =Gene))+
  geom_hline(yintercept = 0)+
  geom_col(position = position_dodge(preserve = "single"),
           colour = "black",
           alpha = 0.5)+
  scale_y_continuous(breaks = seq(500,
                                  23000,
                                  by = 2000))+
  theme_minimal(base_size = 20)+
  theme(plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_blank())+
  theme(legend.position = "none")+
  labs(x = "Genes",
       y = "Total Read Counts",
       title = "A Post-treatment Sample of pCR Patient_03 filterd by 500 > Counts")

figure3 <- p1 / p2

plot(figure3)

ggsave("../results/figure_03.jpg",
       plot = figure3,
       width = 20,
       height = 10,
       device = "jpeg")
```

The 4. figure displays the differences in read counts between the pre-and post-treatment by genes. As an example, we arbitrary chose 1-1 no_pCR-patient from each treatment type and filtered the read counts (Count \> 2500).

```{r, fig.height=10, fig.width=20}
#| echo: false
p3 <- full_reads_data |>
  filter(pCR_status == "no_pCR" &
         subject_ID == "subject_12" &
         Count > 500 &
         Treatment == "pre_treatment") |>
  ggplot(aes(x = Gene,
             y = Count,
             fill = Gene))+
  geom_hline(yintercept = 0)+
  geom_col(position = position_dodge(preserve = "single"),
           colour = "black",
           alpha = 0.5)+
  scale_y_continuous(breaks = seq(500,
                                  10000,
                                  by = 1000))+
  theme_minimal(base_size = 20)+
  theme(plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_text(angle = 60,
                                   vjust = 1,
                                   hjust = 1,
                                   size = 15))+
  theme(legend.position = "none")+
  labs(x ="Genes",
       y = "Total Read Counts",
       title = "A Pre-treatment Sample of no_pCR Patient_12 filterd by 500 > Counts")

p4 <- full_reads_data |>
  filter(pCR_status == "no_pCR" &
         subject_ID == "subject_12" &
         Count > 500 &
         Treatment == "post_treatment") |>
  ggplot(aes(x = Gene,
             y = Count,
             fill = Gene))+
  geom_hline(yintercept = 0)+
  geom_col(position = position_dodge(preserve = "single"),
           colour = "black",
           alpha = 0.5)+
  scale_y_continuous(breaks = seq(500,
                                  10000,
                                  by = 1000))+
  theme_minimal(base_size = 20)+
  theme(plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_text(angle = 60,
                                   vjust = 1,
                                   hjust = 1,
                                   size = 15))+
  theme(legend.position = "none")+
  labs(x ="Genes",
       y = "Total Read Counts",
       title = "A Post-treatment Sample of no_pCR Patient_12 filterd by 500 > Counts")

figure4 <- p3 / p4

plot(figure4)

ggsave("../results/figure_04.jpg",
       plot = figure4,
       width = 20,
       height = 10,
       device = "jpeg")
```

Our last figure illustrates the differences in read counts between the no_pCR and pCR patients in each treatment type by genes. We have to many genes so we also filtered the read counts here.

```{r, fig.height=27, fig.width=27}
#| echo: false
p5 <- full_reads_data |>
  filter(Count > 2500 &
         Treatment == "pre_treatment") |>
  ggplot(aes(x = Gene,
             y = Count,
             fill = pCR_status))+
  geom_boxplot()+
  geom_hline(yintercept = 2000)+
  scale_y_continuous(breaks = seq(2000,
                                  10000000,
                                  by = 1000))+
  theme_minimal(base_size = 25)+
  theme(plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_text(angle = 60,
                                   vjust = 1,
                                   hjust = 1,
                                   size = 20))+
  labs(x ="Genes",
       y = "Total Counts",
       title = "Total Read Counts of the genes before the treatment filtered by pCR_status")

p6 <- full_reads_data |>
  filter(Count > 2500 &
         Treatment == "post_treatment") |>
  ggplot(aes(x = Gene,
             y = Count,
             fill = pCR_status))+
  geom_boxplot()+
  geom_hline(yintercept = 2000)+
  scale_y_continuous(breaks = seq(2000,
                                  10000000,
                                  by = 1000))+
  theme_minimal(base_size = 25)+
  theme(plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        axis.text.x = element_text(angle = 60,
                                   vjust = 1,
                                   hjust = 1,
                                   size = 20))+
  labs(x ="Genes",
       y = "Total Counts",
       title = "Total Read Counts of the genes after the treatment filtered by pCR_status")

figure5 <- p5 / p6

plot(figure5)

ggsave("../results/figure_05.jpg",
       plot = figure5,
       width = 27,
       height = 27,
       device = "jpeg")
```
